+++ database/factories/UserFactory.php
@@ -26,7 +26,7 @@ class UserFactory extends Factory
             'name' => fake()->name(),
             'email' => fake()->unique()->safeEmail(),
             'email_verified_at' => now(),
-            'password' => static::$password ??= Hash::make('password'),
+            'password' => static::$password ??= Hash::make('S3curePass!789'),
             'remember_token' => Str::random(10),
             'status' => fake()->randomElement(UserStatus::cases())->value,
         ];